--- packages/core/src/animation/internal/AnimationCurveOwner/index.ts
@@ -1,1 +1,0 @@
-export { AnimationCurveOwner } from "./AnimationCurveOwner";
