+++ packages/core/src/animation/internal/AnimatorStateData.ts
@@ -1,5 +1,5 @@
 import { KeyframeValueType } from "../Keyframe";
-import { AnimationCurveOwner } from "./AnimationCurveOwner";
+import { AnimationCurveOwner } from "./AnimationCurveOwner/AnimationCurveOwner";
 import { AnimationEventHandler } from "./AnimationEventHandler";
 
 /**
